Feat: 상품조회, 상품수정 테스트 코드 완료 #4

+++ admin_page/src/main/java/com/playground/admin_page/main/product/dto/ProductDto.java
@@ -4,13 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.sql.Blob;
 import java.time.LocalDateTime;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductDto {
-    private int productId;
+    private Long productId;
     private String productName;
     private String category;
     private byte[] productImage;
